Playwright_with_Fixture/Playwright_Project_1/requirements.txt
aiofiles==25.1.0
aiogram==3.24.0
aiohappyeyeballs==2.6.1
aiohttp==3.13.3
aiosignal==1.4.0
allure-pytest==2.15.3
allure-python-commons==2.15.3
annotated-types==0.7.0
attrs==25.4.0
autobahn==25.12.2
cbor2==5.7.1
certifi==2025.11.12
cffi==2.0.0
charset-normalizer==3.4.4
colorama==0.4.6
cryptography==46.0.3
decorator==5.2.1
Faker==39.0.0
frozenlist==1.8.0
greenlet==3.3.0
h11==0.16.0
hyperlink==21.0.0
idna==3.11
iniconfig==2.3.0
magic-filter==1.0.12
msgpack==1.1.2
multidict==6.7.0
outcome==1.3.0.post0
packaging==25.0
playwright==1.57.0
pluggy==1.6.0
propcache==0.4.1
py-ubjson==0.16.1
pycparser==2.23
pydantic==2.12.5
pydantic_core==2.41.5
pyee==13.0.0
Pygments==2.19.2
PySocks==1.7.1
pytest==9.0.2
pytest-asyncio==1.3.0
pytest-base-url==2.1.0
pytest-playwright==0.7.2
pytest-tornasync==0.6.0.post2
pytest-trio==0.8.0
python-slugify==8.0.4
requests==2.32.5
selenium==4.39.0
sniffio==1.3.1
sortedcontainers==2.4.0
text-unidecode==1.3
tornado==6.5.4
trio==0.32.0
trio-websocket==0.12.2
txaio==25.12.2
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.3
ujson==5.11.0
urllib3==2.6.2
websocket-client==1.9.0
wsproto==1.3.2
yarl==1.22.0
